HW2: refactor modules - using directory approach
